AoS and SoA finished inplementation for conveyors
Fixed and updated all the funcions for the SoA of the conveyors
Updated behaviour for the InsertItem of the Miner to adjust to both systems

+++ FactoryWorld/Source/AoSConveyorSystem.cpp
@@ -7,7 +7,7 @@ AoSConveyorSystem::~AoSConveyorSystem()
 	mConveyors.clear();
 }
 
-ConveyorId AoSConveyorSystem::Create(float cX, float cY, GameObject::Orientation o)
+ConveyorId AoSConveyorSystem::Create(int cX, int cY, GameObject::Orientation o)
 {
 	Conveyor* c = new Conveyor(o);
 	GridManager& gridManager = GridManager::GetInstance();
